New web page: widearea_nodeinfo.php3
Intended as a replacement for the old widearea_info page, which
was just not scaling with the number of nodes we now have in the
database.

Allows the user to give a set of criteria (such as node, bandwidth,
latency, and packet loss), and grabs links that match those criteria.