Initial DB support for wireless nodes. Added a "protocol" text field and an "is_accesspoint" boolean to virt_lans. The former defaults to "ethernet" but can be set to anything (80211a, 80211b, etc) in the NS file. The is_accesspoint is temporary, and simply allows you to set which node is the accesspoint in the NS file. This slot will probably move to another table at some point.
Added interface_capabilities table, which is intended to list the
capabilities and the default values, for interfaces listed in the
interface_types table. This allows a more flexible description of
interfaces, expecially wireless devices. Initially, I have seeded the
table with just the default protocol (ethernet) and the speed. For
example, the fxp:

	fxp              | protocols         | ethernet |
	fxp              | ethernet_defspeed | 100000   |

As you can see, protocols is plural, and is intended to be a comma
separated list. So, for the atheros wireless card:

	ath              | protocols         | 80211a,80211b,80211g |
	ath              | 80211a_defspeed   | 54000                |
	ath              | 80211b_defspeed   | 11000                |
	ath              | 80211g_defspeed   | 54000                |

I gave up on using the entire row as a primary key; this is just too
painful from perl/php/python where hashes are the most popular data
structure, and duplicate columns get overwritten.