First cut at client side configuration of wireless nodes. Redhat only, no freebsd support.
The primary change is that tmcd now sends down a list of setting to
apply to each interface, and that list is turned into a hash table,
and provided to rc.config, which passes them along to the machine
dependent routine in liblocsetup.

Then in the linux version of liblocsetup there is a bunch of new code
to configure wireless links using iwconfig and iwpriv, using the
settings array.

All of this is bound to change once we get more experience with it.