Replace old and tired webglimpse with new and modern index/search soeftware called swish-e since webglimpse was broken, did not install out of the box, no longer being maintained by the authors, and generally a pain in the ass.
swish-e is probably overkill, but it works and can be configured and
run from a single configuration file (swish.conf.in).

I've removed the advanced search form and replaced with simple form.
I can bring back advanced stuff if needed, although better to wait
until the swish-e port is upgraded and there is better backend support
for generating the html output.

There is a backend script called websearch that is called from the php
search page. It invokes swish and spits out html that is amenable to
being inbedded in our existing page layout (we were not able to do
that with webglimpse).