Add five new modules:
  imageid, osid - Get listings of image/OS IDs.
  fs - Get directory listings for the emulab file system.
  user - Get information about the user.
  emulab - Get information about emulab as a whole.

Added new method response, RESPONSE_REFUSED, which is returned
whenever emulab is down or otherwise unavailable.

Arguments to methods can now be strings or typed XML-RPC data.  For
example, a boolean parameter will accept (in python) True or the
string "true".

Added osid and status values to the "mapping" aspect of the
"experiment.info" method.  Also added "experiment.constraints" for
getting the idle threshold and "experiment.getlist" for getting the
list of experiments where the user is the head.

Added "node.console" method to get ACL data for connecting to a node's
console.