Change the way we do wireless links.
In the ptop file, put the specific protocol (ie. 80211a) on the link
line.

In the ptop file, there is now only one fake airswitch. It can take
on all wireless types. We now only print one link line for each
physical wireless interface - it has the maximum bandwidth supported
by the interface, and is able to take on any protocol supported by the
interface.