Small hack requested by Jay to aid in creating wireless nodes; added new directive for your NS file: tb-use-physnaming(1) which means that if you have a line like this:
	set pc6 [$ns node]

and pc6 is an actual name of a physical node, then do an implicit
"fixnode" to pc6. In other words, pc6 is the name of the node in your
NS file (virtual topo) *and* will be fixed to the real pc6. Needless
to say, if pc6 is allocated when you go to swapin, you lose.