Silly script to take the testbed commit log that is rsync'ed over each night, and process it into a file to stick into the testbed www directory. This replaces the current approach of wishfully thinking that people will remember to add their commit log messages to another file and them install it. I'm one of the worst offenders these days ...
So, take the commit log, skip up to the first entry in 2004/01, then
take the rest of the file and reverse the entries. Copy the resulting
file over to the www directory.

This script is run from cron on our site only!

Note to Opera users; opera does not respect <XMP> properly, and so the
page is somewhat scrogged by HTML entities that are supposed to be
ignored when inside an XMP. Works okay on other browsers.