Add makefile support for taking a FreeBSD fixit disk and turning that into a testbed boot CD in a mostly automated manner. There are still a few things that need to be done by hand, which are described in the cdboot/README file.
Add tmcd/freebsd/cdboot directory of little scripts and files that
need to onto the fixit disk.

The install target is cdboot-install (must be run as root) and if you
are brave enough to run it, you better give it a DESTDIR argument or
you will write a bunch of files onto the local node that will cause
mayhem and havoc at the next reboot.