Add new usage for doing route calculation on nodes instead of boss:
	staticroutes -j [-n] [-f] [-d] <pid> <eid>

Use -j option to generate a "linkmap" file in the current directory.
The linkmap is of the current format (say, for a lan of three members):

        3 3
	nodea 10.0.1.1 nodeb 10.0.1.2 2
	nodea 10.0.1.1 nodec 10.0.1.3 2
	nodeb 10.0.1.2 nodec 10.0.1.3 2

The last column is the cost metric.

Use -n option to print the linkmap to stdout instead of creating the
file. -f is the usual force option (ignores routertype) and -d prints
some debugging goo to stderr.

For Shashi's 2440 vnode example (2610 edges):

	{26} boss$ time perl ./staticroutes -f -j testbed 2440
	real    0m1.431s
	user    0m0.830s
	sys     0m0.147s