Add "distributed" calculation of the hosts file. If a topomap file is present (gettopmap() returns valid) turn the topomap into input to feed into the genhostsfile program (see initial revision for os/genhostsfile), Take the output and feed that into the parsing code.
NB: If we change the output of tmcd hostnames call, we have to change the
output of this program too. This will likely become a problem at some
point, although the format of the hostnames command has not changed for
about 2 years now.

The point of doing this is to remove the bottleneck of computing hostsfiles
from tmcd, which centrally computes them for every node, and in a big topo
(100s of nodes) that takes a lot of time and bw.