Prep work for DHCP discovery of control net interface:
- have pump/dhclient script record the DHCP'ed interface in
  /var/emulab/boot/control_interface

- change control_interface script to first check for that file
  and use the contents if it exists

Note that, as of this commit, we are still telling pump/dhclient which
interface to DHCP on (i.e., we still determine the control net interface
the old way to invoke pump/dhclient) so this commit is not that useful.

What still has to be done is to change the startup to invoke dhclient/pump
on all interfaces.  This turns out to be a royal, royal PITA.

Stay tuned...