Major re-structuring of the newclient code, and merged it with the writefloppy code. The point of doing this was for headless nodes, like the new wireless nodes - both scripts have similar needs in terms of communicating with the user through beep codes.
Added a '-h' option for headless operation, in which we go into an
infinite loop with a beep code on an error.

Added a '-w' option to make it write the node ID picked by boss to
a floppy.

Added a '-t' option, useful for testing.

Grab the node's CPU speed directly, instead of using cpuspeed.awk -
we don't want to have to keep updating that in the newnode MFS.

At some point, I will probably add the ability to do DHCP to this
script, since that's another thing that could fail that we'd like to
have a beep code for headless nodes.