Back out previous revision (-j); linkmap is now generated on experimental nodes from the topomap.
"static" routing now implies ddijk routing, but since old images won't
have any of that support, we still have to generate routes and store
them in the DB. However, only doing this when the number of nodes in
the experiment is less than 250; otherwise print a warning and refuse
to generate the routes.

Also add "static-old" so that people can request/force old static
routing no matter how big the experiment.