Kluge around problems with picking nodes in HyperViewer on Windows. I will have to fix the bug in PyOpenGL argument passing to get picking working right.
Meanwhile, I put in a simple work-around hack: when you pick a graph
edge in the spanning tree displayed by HyperViewer, the node it leads
to is picked instead.  (The edge has a negative integer ID, while the
node it leads to is positive.)

This also gives the ability to "zip" along edges of the graph, which
I kind of like.  It's not a very good substitute for picking nodes,
but might be worth keeping even after node picking works.  We can
also define our own semantics for picking edges in the graph...