Add three new methods to return all the event system goo to the event scheduler running on ops.
Okay, so this is probably not an ideal mechanism for wrapping up what are
essentially DB queries, since there are no xmlrpc-over-ssh bindings for
C. So we have to run an external client (sshxmlrpc_client) to talk to the
server. This is not so bad in itself, but returning the eventlist as a
single response seems like an unnatural approach for returning what is
essentialy a stream of data. Besides, unless we expect people might want to
write their own schedulers (highly unlikely), exporting this stuff via the
XMLRPC server is a silly thing to do.

Instead, some kind of protected DB access proxy might be more appropriate.
I have no idea how this might operate, but its something to think about.