Fix shell quoting problem which I indirectly introduced just before the big power down when I changed sshtb to not invoke a subshell wrapper, but to exec ssh directly. Built into tbacct was an extra pair of \\ escapes to protect the outer double quotes from that extra subshell.  When I removed that subshell, the extra escapes wreaked havoc.
Needless to say, I really want to change how accounts are built on ops
to use tmcd like a regular experimental node. We can almost do that
now, except for the little detail that sending over 800 users would be
a lot of traffic for single updates. I've been meaning to extend the
protocol to allow for single updates, but have not had time yet!