A couple of minor fixups:
* the python mysql client returns a datetime object rather than a string
  sequence - must convert it to string

* PLAB moved the URL to their programmatic API - adjusted on our end.

* PLAB changed format of expiration time in listSlices() RPC from a
  datetime string to seconds since the epoch in UTC.

There is still an outstanding issue that I have mailed plab support about:
The listSlices() RPC is only returning the first entry in the list of slices
we have allocated at PLC (the same call used to grab all slice entries).
This is causing us to fail on renewal for the unlisted slices.