Several enhancements to the installation scripts:
In ops-install, only make mailing list files for email addresses
that we think will be hosted on ops.

To both boss-install and ops-install, add a '-p' option to specify
a directory in which binary packages can be found - so that we can
avoid building ports.

Fix up the way we handle the elvind config file - it must be identical
to the one in the source tree. Add a library function to check for
this.

These are as yet untested.