Add clause to node query to select just experimental nodes when -e option was given (as is when called from experiment page). This restricts the nodes to just those in the experiment.
This interacts with ghosting mode (where other nodes on other floors
are shown). In the above case, should ghosting show all nodes on other
floors, or just nodes in the experiment on other floors? For now, it
does the latter, but it might useful to have a mode for the other
case?