Remove the SSLCERT defs and the network defs from configure proper. Move them to defs-default.
Allow for per-user defs files to use . defs-default in them. I do this
by temporarly changing the path when evaluating the defs file, adding
the directory where the defs file lives. So, you can do this in your
personal defs file, whic saves on a lot of typing!

	. defs-default
	TBOPSEMAIL=stoller@flux.utah.edu
	TBLOGSEMAIL=stoller@flux.utah.edu
	blah
	blah
	blah

I also changed all of the *local* defs file to use the above format.